[splats,openvdb] backface culling + only 1 voxel thick shell
I've added a dragon here so it's easier to test performance. The
next step is to bucket the splats and vector quantize them as
perf dropped dramatically when I added normals (16 extra bytes per
splat).

By bucketing neighboring splats we can assign the bucket a high
precision position and normal and contained splats can store less data
as it will be more of a delta from the bucket pos/norm.

If a large number (256 maybe) of splats are packed into a bucket we
can speed up the backface culling and rendering by operating on buckets
before reaching into the finer grained splat arrays.