this works if you treat vox bounds as rigid bodies
as soon as you try and add voxels as sub-cubes this
breaks in a variety of ways.

There may be room left in this approach if we hijack the
collision response and do pairwise tests against the
underlying voxel grids, but that seems pretty expensive.
Moving the contact response to the GPU also seems slow
due to having to cross the pci-e bridge twice to get an
answer.

Perhaps porting qu3e to glCompute would help...